+++ CHANGELOG.md
@@ -2,8 +2,23 @@
 
 ## develop
 
-## v0.17.1
+### jannovar-cli
 
+* CLI changes such that one VCF input and one VCF output path can be used only
+* Replacing apache commons-cli with argparse4j for a more modern and usable CLI
+* Consistently writing out HUGO symbols for gene names, using the `hgnc_complete_set.txt` information downloaded when building the annotation DB
+* Upgrading from ENSEMBL-74 to ENSEMBL-75 for annotation database files
+* Removing support for old Jannovar-style annotations (#241)
+
+### jannovar-htsjdk
+
+* Properly annotating Mendelian inheritance for intergenic variants
+
+### jannovar-core
+
+* downloading `hgnc_complete_set.txt` together with data sets, `TranscriptModel` objects now consistently contain additional IDs
+* making ENSEMBL parsing more robust (falling back to transcript name if no transcript ID)
+* fixing bug #248 for ENSEMBL that used `gene_id` for `gene_name`
 * bugfix of NullPointerException in RefSeqParser while parsing refSeq curated
 
 ## v0.17